Use Event Emitter instead of Stream abstraction
Previously, I was using Streams for everything, including parsing the
HTML, totaling the rain, and outputting the content. I don't think that's
the best abstraction here, especially since the totaler needs to wait
until all of the data is downloaded before responding anyway. So, switch
ObservationParser to an Event Emitter pattern (even though it is still a
Writable stream), and have Totaler watch the events.